Fixed #85 moving S7Type from impl to api package.
