+++ src/main/java/com/github/s7connector/blocks/CONT_C.java
@@ -16,7 +16,7 @@ limitations under the License.
 package com.github.s7connector.blocks;
 
 import com.github.s7connector.api.annotation.S7Variable;
-import com.github.s7connector.impl.utils.S7Type;
+import com.github.s7connector.api.S7Type;
 
 /**
  * PID Control block representation
