+++ src/main/java/com/github/s7connector/impl/serializer/converter/BitConverter.java
@@ -16,7 +16,7 @@ limitations under the License.
 package com.github.s7connector.impl.serializer.converter;
 
 import com.github.s7connector.api.S7Serializable;
-import com.github.s7connector.impl.utils.S7Type;
+import com.github.s7connector.api.S7Type;
 
 /**
  * The Class BitConverter is responsible for converting bit values
