+++ src/main/java/com/github/s7connector/impl/serializer/parser/BeanEntry.java
@@ -18,7 +18,7 @@ package com.github.s7connector.impl.serializer.parser;
 import java.lang.reflect.Field;
 
 import com.github.s7connector.api.S7Serializable;
-import com.github.s7connector.impl.utils.S7Type;
+import com.github.s7connector.api.S7Type;
 
 /**
  * A Bean-Entry
