+++ src/test/java/com/github/s7connector/test/example/MyDataBean.java
@@ -16,7 +16,7 @@ limitations under the License.
 package com.github.s7connector.test.example;
 
 import com.github.s7connector.api.annotation.S7Variable;
-import com.github.s7connector.impl.utils.S7Type;
+import com.github.s7connector.api.S7Type;
 
 /**
  * @author Thomas Rudin (thomas@rudin-informatik.ch)
